Develop to Main 브랜치 최종 병합

+++ src/main/java/com/sparta/newsfeed_project/domain/buddies/service/BuddiesService.java
@@ -1,15 +1,15 @@
 package com.sparta.newsfeed_project.domain.buddies.service;
 
 
-import com.sparta.newsfeed_project.domain.buddies.dto.RequestBuddiesDto;
 import com.sparta.newsfeed_project.domain.buddies.dto.ResponseBuddiesDto;
 import com.sparta.newsfeed_project.domain.buddies.entity.Buddies;
 import com.sparta.newsfeed_project.domain.buddies.repository.BuddiesRepository;
+import com.sparta.newsfeed_project.domain.common.exception.ResponseCode;
+import com.sparta.newsfeed_project.domain.common.exception.ResponseException;
+import com.sparta.newsfeed_project.domain.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -17,41 +17,89 @@ import java.util.List;
 public class BuddiesService {
 
     private final BuddiesRepository buddiesRepository;
+    private final MemberRepository memberRepository;
 
-    public ResponseBuddiesDto createBuddies(RequestBuddiesDto requestBuddiesDto) {
-        if(requestBuddiesDto.getFromUserId().equals(requestBuddiesDto.getToUserId())){
-            throw new IllegalArgumentException("중복된 아이디입니다.");
+
+    /**
+     * 회원의 친구 관계를 등록합니다
+     *
+     * @param memberId,userId 회원 ID,친구 ID
+     * @Return 등록된 친구 관계
+     * @since 2024-10-23
+     */
+    public ResponseBuddiesDto createBuddies(Long memberId, Long userId) throws ResponseException {
+        if (!memberRepository.existsById(userId)) {
+            throw new ResponseException(ResponseCode.MEMBER_NOT_FOUND);
         }
-        if(buddiesRepository.existsByFromUserIdAndToUserId(requestBuddiesDto.getFromUserId(), requestBuddiesDto.getToUserId())){
-            throw new IllegalArgumentException("존재하는 친구 목록입니다.");
+        if (memberId.equals(userId)) {
+            throw new ResponseException(ResponseCode.BUDDIES_DUOLICATION_ID_ERROR);
         }
-        Buddies buddies = Buddies.from(requestBuddiesDto);
-        buddies.Approved(true);
+        if (buddiesRepository.existsByFromUserIdAndToUserId(memberId, userId)) {
+            throw new ResponseException(ResponseCode.BUDDIES_DUOLICATION_ERROR);
+        }
+        Buddies buddies = Buddies.from(memberId, userId);
+        buddies.approved(true);
         buddiesRepository.save(buddies);
-        Buddies acceptBuddies = Buddies.upend(requestBuddiesDto);
+        Buddies acceptBuddies = Buddies.upend(userId, memberId);
         buddiesRepository.save(acceptBuddies);
         return buddies.to();
     }
 
-    public List<ResponseBuddiesDto> getBuddies() {
+    /**
+     * 모든 친구 관계의 목록을 불러옵니다
+     *
+     * @return 친구 목록
+     */
+    public List<ResponseBuddiesDto> getBuddies() throws ResponseException {
         List<Buddies> buddies = buddiesRepository.findAll();
+        if (buddies.isEmpty()) {
+            throw new ResponseException(ResponseCode.BUDDIES_NULL_ERROR);
+        }
         return buddies.stream().map(Buddies::to).toList();
     }
 
-    public void acceptBuddies(RequestBuddiesDto requestBuddiesDto) {
+    /**
+     * 회원의 친구 관계를 수락합니다
+     *
+     * @param memberId,userId 회원 ID,친구 ID
+     * @since 2024-10-23
+     */
+    public void acceptBuddies(Long memberId, Long userId) throws ResponseException {
+        if (!buddiesRepository.existsById(userId)) {
+            throw new ResponseException(ResponseCode.BUDDIES_DUOLICATION_ID_ERROR);
+        }
         Buddies buddies = buddiesRepository
-                .findOneByFromUserIdAndToUserId(requestBuddiesDto.getFromUserId(),
-                        requestBuddiesDto.getToUserId());
-        buddies.Approved(true);
+                .findOneByFromUserIdAndToUserId(memberId, userId);
+        buddies.approved(true);
         buddiesRepository.save(buddies);
     }
 
-    public void deleteBuddies(RequestBuddiesDto requestBuddiesDto) {
+    /**
+     * 회원의 친구 관계를 취소합니다
+     * 둘다 관계를 취소했을때 목록에서 삭제합니다
+     *
+     * @param memberId,userId 회원 ID,친구 ID
+     * @since 2024-10-23
+     */
+    public void deleteBuddies(Long memberId, Long userId) {
         Buddies buddies = buddiesRepository
-                .findOneByFromUserIdAndToUserId(requestBuddiesDto.getFromUserId(), requestBuddiesDto.getToUserId());
-        buddies.Approved(false);
+                .findOneByFromUserIdAndToUserId(memberId, userId);
+        if (buddies == null) {
+            throw new ResponseException(ResponseCode.BUDDIES_NULL_ERROR);
+        }
+
         Buddies buddy = buddiesRepository
-                .findOneByFromUserIdAndToUserId(requestBuddiesDto.getToUserId(), requestBuddiesDto.getFromUserId());
+                .findOneByFromUserIdAndToUserId(userId, memberId);
+        if (buddy == null) {
+            throw new ResponseException(ResponseCode.BUDDIES_NULL_ERROR);
+        }
+
+        if (!buddies.isApproved() || !buddy.isApproved()) {
+            throw new ResponseException(ResponseCode.BUDDIES_NULL_ERROR);
+        }
+
+        buddies.approved(false);
+
         if (buddies.isApproved() == buddy.isApproved()) {
             buddiesRepository.delete(buddy);
             buddiesRepository.delete(buddies);
@@ -60,17 +108,18 @@ public class BuddiesService {
         }
     }
 
+    /**
+     * 회원의 친구 관계를 모두 불러옵니다
+     *
+     * @param memberId 회원 ID
+     * @return 친구 목록
+     */
     public List<ResponseBuddiesDto> getAllBuddies(Long memberId) {
         List<Buddies> buddies = buddiesRepository.findAllByToUserId(memberId);
+        if (buddies.isEmpty()) {
+            throw new ResponseException(ResponseCode.BUDDIES_NULL_ERROR);
+        }
         return buddies.stream().map(Buddies::to).toList();
     }
-
-
-    public void deletedBuddies(Long memberId) {
-        Buddies fromUser= buddiesRepository.findByFromUserID(memberId);
-        Buddies toUser = buddiesRepository.findByFromUserID(fromUser.getToUserId());
-        buddiesRepository.delete(fromUser);
-        buddiesRepository.delete(toUser);
-    }
-
 }
+
